feat(asset-reminders): create function to resolve reminders actions
